# backend/app/Models/Category.php
<?php
// app/Models/Category.php
namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Category extends Model
{
    use HasFactory;

    protected $fillable = ['name_ar', 'name_fr'];

    public function reports()
    {
        return $this->hasMany(Report::class);
    }
}