# backend/app/Virtual/Models/Category.php
<?php

namespace App\Virtual\Models;

/**
 * @OA\Schema(
 *     title="Category",
 *     description="Category model",
 *     @OA\Xml(
 *         name="Category"
 *     )
 * )
 */
class Category
{
    /**
     * @OA\Property(
     *     title="ID",
     *     description="ID of the category",
     *     format="int64",
     *     example=1
     * )
     *
     * @var integer
     */
    private $id;

    /**
     * @OA\Property(
     *     title="Name (Arabic)",
     *     description="Name of the category in Arabic",
     *     example="صيانة"
     * )
     *
     * @var string
     */
    private $name_ar;

    /**
     * @OA\Property(
     *     title="Name (French)",
     *     description="Name of the category in French",
     *     example="Maintenance"
     * )
     *
     * @var string
     */
    private $name_fr;

    /**
     * @OA\Property(
     *     title="Created at",
     *     description="Creation date",
     *     example="2025-05-20T12:00:00.000000Z",
     *     format="datetime",
     *     type="string"
     * )
     *
     * @var \DateTime
     */
    private $created_at;

    /**
     * @OA\Property(
     *     title="Updated at",
     *     description="Last update date",
     *     example="2025-05-20T12:00:00.000000Z",
     *     format="datetime",
     *     type="string"
     * )
     *
     * @var \DateTime
     */
    private $updated_at;
}